Extract and re-use copy-pasted beforeUnloadHandler to useWarnOnUnsavedChanges.ts
